UT fix

+++ libs/features/mas/web-components/test/merch-card.mini-compare.test.html.js
@@ -52,7 +52,7 @@ runTests(async () => {
                         const el =
                             miniCompareChart.shadowRoot.querySelector(selector);
                         if (!el) return 0;
-                        return parseFloat(window.getComputedStyle(el).height);
+                        return parseInt(window.getComputedStyle(el).height);
                     })
                     .join(',');
                 return heights;
@@ -76,7 +76,7 @@ runTests(async () => {
                     .fill()
                     .map(
                         (_, i) =>
-                            parseFloat(
+                            parseInt(
                                 window.getComputedStyle(
                                     miniCompareChart.querySelector(
                                         `.footer-row-cell:nth-child(${i + 1})`,
